세션 래퍼 함수 작성 및 기존 코드 교체
