+++ server/api/chat/index.post.ts
@@ -1,5 +1,5 @@
 import prisma from "~/lib/prisma";
-import { GET_SESSION_CONFIG } from "~/constants/session-const";
+import { withGetSession } from "~/lib/withSession";
 
 export interface CreateChatResponse {
   ok: boolean;
@@ -25,7 +25,7 @@ export default defineEventHandler(async (event) => {
   try {
     const {
       data: { userId },
-    } = await getSession(event, GET_SESSION_CONFIG);
+    } = await withGetSession(event);
 
     const chat = await createChat(userId);
